==> 4_orders.ipynb <==
import credentials as crs
from fyers_apiv3 import fyersModel
import pandas as pd
import datetime as dt
#credentials
# Replace these values with your actual API credentials
client_id = crs.client_id
secret_key = crs.secret_key
redirect_uri = crs.redirect_uri
with open('access.txt') as f:
    access_token=f.read()

# Initialize the FyersModel instance with your client_id, access_token, and enable async mode
fyers = fyersModel.FyersModel(client_id=client_id, token=access_token,is_async=False, log_path="")


# market order

data={ "symbol": "NSE:HATHWAY-EQ", 
      "qty": 1, 
      "type": 2, 
      "side": 1, 
      "productType": "INTRADAY", 
      "limitPrice": 0, 
      "stopPrice": 0, 
      "validity": "DAY", 
      "stopLoss": 0, 
      "takeProfit": 0, 
      "offlineOrder": False, 
      "disclosedQty": 0 }
response1 = fyers.place_order(data=data)
print(response1)

# cnc order


data={ "symbol": "NSE:YESBANK-EQ", 
      "qty": 1, 
      "type": 2, 
      "side": 1, 
      "productType": "CNC", 
      "limitPrice": 0, 
      "stopPrice": 0, 
      "validity": "DAY", 
      "stopLoss": 0, 
      "takeProfit": 0, 
      "offlineOrder": False, 
      "disclosedQty": 0 }
response1 = fyers.place_order(data=data)
print(response1)

# limit order

data = {
    "symbol":"NSE:YESBANK-EQ",
    "qty":1,
    "type":1,
    "side":1,
    "productType":"INTRADAY",
    "limitPrice":20,
    "stopPrice":0,
    "validity":"DAY",
    "disclosedQty":0,
    "offlineOrder":False,
    "orderTag":"tag1"
}
response = fyers.place_order(data=data)
print(response)

# stop order market

data = {
    "symbol":"NSE:HATHWAY-EQ",
    "qty":1,
    "type":3,
    "side":-1,
    "productType":"INTRADAY",
    "limitPrice":0,
    "stopPrice":22,
    "validity":"DAY",
    "disclosedQty":0,
    "offlineOrder":False,
    "orderTag":"tag1"
}
response3 = fyers.place_order(data=data)
print(response3)

# stop limit order


data = {
    "symbol":"NSE:HATHWAY-EQ",
    "qty":1,
    "type":4,
    "side":-1,
    "productType":"INTRADAY",
    "limitPrice":21,
    "stopPrice":22,
    "validity":"DAY",
    "disclosedQty":0,
    "offlineOrder":False,
    "orderTag":"tag1"
}
response3 = fyers.place_order(data=data)
print(response3)

# cover order


data = {
    "symbol":"NSE:ONGC-EQ",
    "qty":1,
    "type":2,
    "side":1,
    "productType":"CO",
    "limitPrice":0,
    "stopPrice":0,
    "validity":"DAY",
    "disclosedQty":0,
    "offlineOrder":False,
    "stopLoss":3,
    "takeProfit":0
}
response3 = fyers.place_order(data=data)
print(response3)

# bracket order


data = {
    "symbol":"NSE:NTPC-EQ",
    "qty":1,
    "type":2,
    "side":1,
    "productType":"BO",
    "limitPrice":0,
    "stopPrice":0,
    "validity":"DAY",
    "disclosedQty":0,
    "offlineOrder":False,
    "stopLoss":5,
    "takeProfit":5
}
response3 = fyers.place_order(data=data)
print(response3)

# braket order with limit price


data = {
    "symbol":"NSE:NTPC-EQ",
    "qty":1,
    "type":1,
    "side":1,
    "productType":"BO",
    "limitPrice":335,
    "stopPrice":0,
    "validity":"DAY",
    "disclosedQty":0,
    "offlineOrder":False,
    "stopLoss":5,
    "takeProfit":5
}
response6 = fyers.place_order(data=data)
print(response3)

# modify order

orderId = "24100300574539"
data = {
    "id":orderId, 
    "type":1, 
    "limitPrice": 332, 
    "qty":1
}

response = fyers.modify_order(data=data)
print(response)

# cancel order

data = {"id":'24022300382793-BO-1'}

responsel = fyers.cancel_order(data=data)
print(responsel)
